Update basicMove (add static fuctions), spyPost (return y and spyPostType) and reposition (int precision)

+++ Inc/middleware/controls/mazeControl/reposition.h
@@ -26,8 +26,8 @@ enum telemeters_used
 void repositionSetInitialPosition(double initial_position);
 void repositionResetTelemeterUsed(void);
 enum telemeters_used repositionGetTelemeterUsed(void);
-double repositionGetFrontDist(void);
+int  repositionGetFrontDist(void);
 void repositionGetFrontDistCal(void);
-int frontCal(float max_speed);
+int  frontCal(float max_speed);
 
 #endif
